Bug 1545864 - stop publishing docs
