--- libraries/docs/test/docs/example.md
@@ -1,6 +1,0 @@
----
-title: Just saying
-description: Nothing to see here
----
-
-This is some text about some stuff that happened sometime ago
