--- libraries/docs/test/docs/nested/nested-example.md
@@ -1,6 +1,0 @@
----
-title: This is Nested
-description: Nested to see here
----
-
-This is some text about some stuff that happened sometime ago, but nested this time.
